+++ occurrences/consts.py
@@ -1,4 +1,4 @@
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 NOTIFICATION_TYPE_EMAIL = "email"
 NOTIFICATION_TYPE_SMS = "sms"