+++ organisations/models.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.db import models
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from helusers.models import AbstractUser
 from organisations.services import (
     send_myprofile_creation_notification_to_admins,