+++ palvelutarjotin/oidc.py
@@ -11,6 +11,22 @@ class GraphQLApiTokenAuthentication(ApiTokenAuthentication):
     2. return only the user object, or django's auth middleware will fail
     """
 
+    def __convert_amr_to_list(self, id_token):
+        """
+        OIDC's amr validation fails, since Tunnistamo sends the amr as a string
+        instead of a list:
+        https://github.com/City-of-Helsinki/tunnistamo/commit/a1b434bbbff92466a144f914a98985008d0ea836.
+
+        To fix the claim validation issue, convert the id_token amr to list
+        when a string is given.
+        """
+        if id_token["amr"] and not isinstance(id_token["amr"], list):
+            id_token["amr"] = [id_token["amr"]]
+
+    def validate_claims(self, id_token):
+        self.__convert_amr_to_list(id_token)
+        super().validate_claims(id_token)
+
     def authenticate(self, request, **kwargs):
         user_auth_tuple = super().authenticate(request)
         if not user_auth_tuple: